+++ lib/src/functions/device_manager/device_manager_provider.dart
@@ -1,11 +1,11 @@
 import 'package:jnap/logger.dart';
 import 'package:jnap/src/functions/device_manager/device_manager_state.dart';
+import 'package:jnap/src/functions/device_manager/devices_extensions.dart';
 import 'package:jnap/src/functions/device_manager/service/backhaul_info_service.dart';
 import 'package:jnap/src/functions/device_manager/service/devices_service.dart';
 import 'package:jnap/src/functions/device_manager/service/network_connections_service.dart';
 import 'package:jnap/src/functions/device_manager/service/radio_info_service.dart';
 import 'package:jnap/src/functions/device_manager/service/wan_status_service.dart';
-import 'package:jnap/src/functions/polling/polling_provider.dart';
 import 'package:jnap/src/functions/polling/providers.dart';
 import 'package:jnap/src/models/jnap_data/back_haul_info.dart';
 import 'package:jnap/src/models/jnap_data/layer2_connection.dart';
@@ -21,59 +21,57 @@ final deviceManagerProvider =
 
 class DeviceManagerNotifier extends Notifier<DeviceManagerState> {
   @override
-  DeviceManagerState build() {
-    final coreTransactionData = ref.watch(pollingProvider).value;
-    return createState(pollingResult: coreTransactionData);
-  }
+  DeviceManagerState build() => DeviceManagerState();
 
   void init() {
     state = DeviceManagerState();
   }
 
-  DeviceManagerState createState({CoreTransactionData? pollingResult}) {
+  Future<void> fetch() async {
     // Init services
+    final radioInfoService = RadioInfoService(ref);
+    final backhaulInfoService = BackhaulInfoService(ref);
+    final wanStatusService = WANStatusService(ref);
     final networkConnectionsService = NetworkConnectionsService(ref);
     final devicesService = DevicesService(ref);
-    final wanStatusService = WANStatusService(ref);
-    final backhaulInfoService = BackhaulInfoService(ref);
-    final radioInfoService = RadioInfoService(ref);
 
-    // Fetch cache data
+    // Polling last update time
     final lastUpdate = ref.read(pollingProvider).value?.lastUpdate;
-    final getDevicesData = devicesService.getDevicesFromCache();
-    final getWANStatusData = wanStatusService.getWANStatusFromCache();
-    final List<Layer2Connection> connectionData =
-        networkConnectionsService.getNetworkConnectionsFromCache();
-    final radioInfoMap = radioInfoService.getRadioInfoMapFromCache();
+    // Radio info
+    final radioInfoMap = await radioInfoService.getRadioInfoMap();
     final guestRadioSettings =
-        radioInfoService.getGuestRadioSettingsFromCache();
-
-    // Start building state
-    var newState = const DeviceManagerState();
+        await radioInfoService.getGuestRadioSettingsData();
     // Backhaul Info
     List<BackHaulInfoData> backhaulInfoList =
-        backhaulInfoService.getBackhaulInfoList();
+        await backhaulInfoService.getBackhaulInfoList();
     // WAN Status
-    RouterWANStatus? wanStatus =
-        wanStatusService.getWANStatusModel(getWANStatusData);
+    RouterWANStatus wanStatus = await wanStatusService.getWANStatusData();
     // Wireless Connections
+    final List<Layer2Connection> connectionsList =
+        await networkConnectionsService.getNetworkConnectionsList();
     Map<String, WirelessConnection> wirelessConnections =
-        networkConnectionsService.getWirelessConnections(connectionData);
+        networkConnectionsService.getWirelessConnectionsMap(
+            connectionsList: connectionsList);
     wirelessConnections =
         networkConnectionsService.updateWirelessConnectionsWithBackhaulInfo(
-            wirelessConnections: wirelessConnections,
-            backhaulInfoList: backhaulInfoList);
+      wirelessConnections: wirelessConnections,
+      backhaulInfoList: backhaulInfoList,
+    );
     // The data process of NetworkConnections MUST be done before building device list
     // Device List
+    final getDevicesData = (await devicesService.getDevices()).output;
     List<LinksysDevice> deviceList = devicesService.getDeviceListAndLocations(
         getDevicesData, wirelessConnections);
-    deviceList = devicesService.updateDeviceIpWithBackhaulInfo(
-        deviceList: deviceList, backhaulInfoList: backhaulInfoList);
-    deviceList =
-        devicesService.updateDeviceWirelessSignalWithWirelessConnections(
-            deviceList: deviceList, wirelessConnections: wirelessConnections);
-    // For getting the master device, update the state
-    newState = newState.copyWith(
+    deviceList
+        .updateIpWithBackhaulInfo(devicesService, backhaulInfoList)
+        .updateWirelessSignal(devicesService, wirelessConnections)
+        .updateUpstream(
+            devicesService,
+            deviceList.firstWhere((device) => device.isMasterDevice),
+            backhaulInfoList);
+
+    // Update state
+    state = state.copyWith(
       wirelessConnections: wirelessConnections,
       deviceList: deviceList,
       wanStatus: wanStatus,
@@ -82,17 +80,6 @@ class DeviceManagerNotifier extends Notifier<DeviceManagerState> {
       backhaulInfoData: backhaulInfoList,
       lastUpdateTime: lastUpdate,
     );
-    // Update device upstream
-    deviceList = devicesService.updateDeviceUpstream(
-        deviceList: deviceList,
-        masterDevice: newState.masterDevice,
-        backhaulInfoList: backhaulInfoList);
-
-    // Update the state with latest device list
-    newState = newState.copyWith(
-      deviceList: deviceList,
-    );
-    return newState;
   }
 
   // Used in cases where the watched DeviceManager is still empty at very beginning stage